% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renderMaps.R
\name{render_map}
\alias{render_map}
\title{Render the Maps for Your Sites}
\usage{
render_map(sites = NULL, siteID = NULL, county)
}
\arguments{
\item{sites}{a path to a semi-colon separated csv file with sites with transect or
point coordinates in WKT format}

\item{siteID}{optional; site id for the sites you want. If not given it use all id in
the `sites` file}

\item{county}{character; the county name of the county your survey is situated in}
}
\value{
a pdf with an overview map and zoomed in maps on the transects or points.
}
\description{
This function takes a list of sites with coordinates of transects, or points and makes
maps from this.
}
\examples{
\dontrun{
render_map("data/lokaler.csv")
}
}
